Bug: Cause a Triple Fault
Guard page is a special memory page at the bottom of a stack that makes
it possible to detect stack overflows. The page is not mapped to any
physical frame, so accessing it causes a page fault instead of silently
corrupting other memory. The bootloader sets up a guard page for our
kernel stack, so a stack overflow causes a page fault.

When a page fault occurs the CPU looks up the page fault handler in the
IDT and tries to push the interrupt stack frame onto the stack. However,
the current stack pointer still points to the non-present guard page.
Thus, a second page fault occurs, which causes a double fault (according
to the above table).

So the CPU tries to call the double fault handler now. However, on a
double fault the CPU tries to push the exception stack frame, too. The
stack pointer still points to the guard page, so a third page fault
occurs, which causes a triple fault and a system reboot. So our current
double fault handler can’t avoid a triple fault in this case.