Print a dot on each timer interrupt
When we print a dot there is a chance that the tests will fail. As when
we do this we have added a basic form of concurrency to our kernel. Even
though Rust helps a lot with concurrency related bugs one noteable
exception is deadlock.

The _start function has the writer lock the interrupt handler waits on
it but _start can only let go of the lock when it finishes but it waits
on the interrupt handler, resulting in a deadlock.