Very basic ability to map a Page to a PageFrame
A page is essentially the starting VirtualAddress and a PageFrame is a
starting PhysicalAddress. Assuming 4KiB size we map a Page to a
PageFrame (a physical location on the memory).

The support is only for 4KiB pages for now. If there is any need of
allocation of a new PageFrame while performing this operation we don't
support that. The behaviour is undefined at this point.